+++ EssentialFeed/EssentialFeedTests/Feed Cache/CacheFeedImageDataUseCaseTests.swift
@@ -1,0 +1,42 @@
+//
+//  CacheFeedImageDataUseCaseTests.swift
+//  EssentialFeedTests
+//
+//  Created by Alex Tapia on 24/04/22.
+//
+
+import XCTest
+import EssentialFeed
+
+class CacheFeedImageDataUseCaseTests: XCTestCase {
+  
+  func test_init_doesNotMessageStoreUponCreation() {
+    let (_, store) = makeSUT()
+    
+    XCTAssertTrue(store.receivedMessages.isEmpty)
+  }
+  
+  func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
+    let (sut, store) = makeSUT()
+    let url = anyURL()
+    let data = anyData()
+    
+    sut.save(data, for: url) { _ in }
+    
+    XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
+  }
+  
+  // MARK: - Helpers
+  
+  private func makeSUT(currentDate: @escaping () -> Date = Date.init,
+                       file: StaticString = #filePath,
+                       line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+    let store = FeedImageDataStoreSpy()
+    let sut = LocalFeedImageDataLoader(store: store)
+    trackForMemoryLeaks(store, file: file, line: line)
+    trackForMemoryLeaks(sut, file: file, line: line)
+    return (sut, store)
+  }
+}
+
+
