+++ EssentialFeed/EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -79,16 +79,6 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
     XCTAssertTrue(received.isEmpty, "Expected no received results after instance has been deallocated")
   }
   
-  func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
-    let (sut, store) = makeSUT()
-    let url = anyURL()
-    let data = anyData()
-    
-    sut.save(data, for: url) { _ in }
-    
-    XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
-  }
-  
   // MARK: - Helpers
   
   private func makeSUT(currentDate: @escaping () -> Date = Date.init,
